pyproject.toml
[project]

name = "sttc"
version = "0.1.0"
description = "sttc - speech to text clipboard - Hotkey-driven audio capture and transcription utility"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = [
    "speech to text clipboard",
    "stt",
    "voice",
    "speech-to-text",
    "clipboard",
    "hotkey",
    "transcription",
    "productivity",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich-click",
    "tqdm",
    "pynput>=1.7.6",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "faster-whisper>=1.0.0",
    "pyperclip>=1.8.2",
    "litellm>=1.42.12",
]

[project.scripts]
sttc = "sttc.cli:main"

[project.optional-dependencies]
extras = ["ipykernel>=7.1.0"]
gui = ["PySide6>=6.8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sttc"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=9.0.2",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "pyinstaller>=6.0.0",
    "types-tqdm",
    "pillow>=11.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",       # docstrings
    "ANN",     # type annotations enforcement
    "TD",      # todo style
    "FIX",     # fixme comments
    "COM",     # trailing commas
    "E501",    # line length
    "DJ",      # django
    "EXE",     # shebang
    "T10",     # debugger
    "TRY003",  # long exception msgs
    "EM101",   # string in exception
    "EM102",   # fstring in exception
    "PLR0913", # too many args
    "PLR2004", # magic values
    "S101",    # assert
    "T20",     # print statements
    "FBT",     # boolean trap
    "RET504",  # unnecessary assignment
    "G004",    # logging fstring
    "BLE001",  # blind exception
    "TRY300",  # try else
    "S311",    # random cryptographic
    "C901",    # too complex
    "PLR0911", # too many return statements
    "PLR0912", # too many branches
    "PLR0915", # too many statements
    "S112",    # logging exception
    "S110",    # try except pass
]
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG001", "ARG002", "PLR2004", "SLF001"]
"**/__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["sttc"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
# strict = true
# warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
disable_error_code = ["import-untyped"]
packages = ["sttc"]

[[tool.mypy.overrides]]
module = [
    "pyperclip",
    "pynput",
    "pynput.*",
    "sounddevice",
    "soundfile",
    "faster_whisper",
    "PySide6",
    "PySide6.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
